requirements.txt
pillow~=10.2.0
PyYAML~=6.0.1
numpy~=1.24.4
moviepy~=1.0.3